repos/payment_gateway/pyproject.toml
[tool.poetry]
authors = ["Engineering <engineering@malaa.tech>"]
description = ""
name = "payment_gateway"
version = "0.1.0"

[tool.poetry.dependencies]
python = "~3.11"
python-dateutil = "~2.8.2"
uvicorn = "^0.27.0.post1"
fastapi = "^0.109.2"
alembic = "^1.13.1"
sqlalchemy-cockroachdb = "^2.0.2"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "23.3.0"
faker = "*"
locust = "^2.14.0"
pylint = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"
pytest-mock = "^3.12.0"
poethepoet = "^0.24.3"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poe.tasks]
api_service = { shell = 'export APP__ROOT_PATH=$(pwd) && python3 -m api.main' }


[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = 'auto'
